+++ Projects/FireAdministrator/Modules/GroupControllerModule/PumpStation/ViewModels/PumpStationDetailsViewModel.cs
@@ -64,8 +64,8 @@ namespace GKModule.ViewModels
 			}
 		}
 
-		int _no;
-		public int No
+		ushort _no;
+		public ushort No
 		{
 			get { return _no; }
 			set
